Move the Shared annotation to the annotations package
