+++ common/src/main/java/com/ably/tracking/common/MessageModels.kt
@@ -1,6 +1,6 @@
 package com.ably.tracking.common
 
-import com.ably.tracking.Shared
+import com.ably.tracking.annotations.Shared
 import com.ably.tracking.GeoJsonMessage
 import com.ably.tracking.Resolution
 import com.google.gson.annotations.SerializedName
